Moved number of functions to utils.py - removed silly docstring associated with /dev/null and Monty Python!
